requirements.txt
flask
google-cloud-bigtable